requirements.txt
aiy-projects-python
json_database
ovos-bus-client
ovos-plugin-manager
